Project Setup
Just general setup for how most of these work. This is the "template" that I am currently using.

I don't use a static template every time I make an app, I simply borrow the changes and additions from the previous most recent project, and adapt them to the needs of the current project. I think it helps with progressing the "general template" I use forward a bit too, because I don't need to necessarily backport the updates to the old project, in order to use them in the new project. I like having it be defined and in concept independently.

Mammoth! 🦣🌨️

I may merge this project into LogLog once this and or that one is more fleshed out. From the consensus of the original idea for this "startup", I think the name was originally going to be "LogLog", and I already have a project going under that name. So it was made to be.