Tabbed Setup
Got some insight from GPT. Much clearer of an abstraction, and more maintainable. The concept was for the stack here to work similar to the Clock app on iOS, i.e. having a tab selection panel, and different panes which are visible only when that tab is selected. Styling and the views' content is for after this step. Wanted to make sure it was inside of a nice foundation before making everything a certain way.