Grid Display
This is the grid-based alternative to the same layout. I do like this implementation because it namely will allow for me to re-sort the display ordering of the elements, while keeping the tab order the same (I want the nav to be tabbable before the main, but the nav to be displayed at the bottom), and GPT recommended if I were to add a sidebar, grid allows for it much easier than flexbox does.

Once I am able to flip the ordering of where main is displayed, then the grid-template-rows part likely would define the 1fr part to be in the middle, rather than the end. Right now, since I haven't though, this is the "correct" option for the time being.