Header + Main Modules
Okay now these are in their own modules, the state is explicitly scoped to the component parameters too. Nice!

Sometimes I feel this makes the declarations a bit more complex than I would like, but it in trade brings more composability to the component internals themselves. Essentially, it allows it to be much more interchangeable, reusable, and modular. That is much more worth it than "small code". I still have to remind myself of this at times. If you don't catch it early enough in your code, it bites back.