Reactive Local Storage
Looking into how I can implement this, as I want to have it so that Local Storage is just as easy to use and bind to as signals are. I wonder if it makes sense to do just that, instead? Have it so that a signal is bound to Local Storage by effects, rather than stores?

https://stackoverflow.com/questions/29971898/how-to-create-an-accurate-timer-in-javascript
https://www.solidjs.com/examples/todos
https://www.solidjs.com/tutorial/stores_createstore