Storage Import Typings
Marked the types with the type descriptor. I like that it makes things explicit.

I am also one that very much enjoys explicit type annotations, as it conveys intent. I think it is great for simple functions, as well as just making it easier to see how things are meant to be seen again at later times when coming back to get familiar with the code again. I am also just used to doing it now because of my time using Rust.