General Nice and Simple Formatting
The code kind of just gets ugly with time, so it's good to see places where you can fix it some more again, before it gets too crazy.

I am trying not to do this too much though, as you have to move fast in a way, to have big ideas to iterate on. Getting all of the small parts perfect to start with only slows you down for actually completing the whole project. I do struggle with that at times still.