Time Clock State
With help from a few online resources, as well as my own reflections and vision, and by help of diving further into things with GPT, this is where we are at.

I really wanted something native to Solid, and now to be fully embedded inside of a component. I wanted it to be it's own state hook. I still feel it's a bit bloated, but I do think it seems to expose what facets I was hoping to make into a single state setup. It handles Local Storage syncing, everything is composed of simple function calls (the user-facing part), and because it is using a store, the raw log values are also reactive. Amazing. Yet to see whether this works, as well as whether it works like I think it should, and whether how I think it should work, should be how it works. These are the dilemmas.

https://stackoverflow.com/questions/29971898/how-to-create-an-accurate-timer-in-javascript
https://www.solidjs.com/examples/todos