Time Clock State Application
Okay, now the new state hook is used in the app! It is not implemented within the UI yet though, that is the next step.

This is great though! This was the first time of realizing how to design the "backend paradigm" for the frontent, and vice versa, and they actually tie together nicely, because they were intentionally designed to work together. Most of my projects end up carving away at both sides of the struggle so they can interweave somehow, versus being like nice, intentional puzzle pieces.