Main Scrolling
Yay, this was getting confusing. It was the `min-height: 0` on main which was needed. I thought it was the children of main which the problem was with. But instead it was because the root grid dimensions weren't able to be passed down to those children because of main not propagating that info.