Log Dates Display
Not pretty yet really, this will get better. Might use a table or something like that. This is meant to be more like a spreadsheet of some sorts.

I will add a way to download everything as JSON too.

I need to add a way to remove the data too, to start fresh. To flush the database, if we were to make puns like that. Not sure who would do such a thing though.