Table Display
Trying to work out what the best ways to display this are, as there is a bunch of metrics that would be nice to see from here. You can extrapolate any timestamps and summaries from it that you'd want, it's still just the same set of points under the hood.