Partial Log Styling
I absolutely love "Recreational", listening back to it again. I very much want to make music for fun again now. Not that that has changed, but I once again see where my place in all of this is. It's becoming clearer. That may even involve burning CDs for people, got a bunch for a super good deal at the thrift store. Really cheap, and awesome to create copies for friends with ease. I love that. Maybe even could mail them to people, for fun. Could hand-write on each one too, like the album and my name. Maybe a personalized note and or their name too, that would be super cool. ~~Might~~ *Will* do it.