Minimum Wage Placeholder
Rather than it being a default value, it is the placeholder text. I plan to make it so that you have to put in your wage to start the timer. That's the main goal of this app, is to track how much you are earning with your rest stops.

This was an idea that was thought of for fun, as part of a group. Yes I know it's weird, that's why it is amusing to me, haha.