setWage Nullish Handling
Okay, now this works properly, from the looks of it.

Terria is awesome, just got into it again today. I think it is one of the first few times I listened all the way through. I have already loved a lot of the songs. I think you have to be in a similar headspace to it to fully feel it though, hence maybe why I've had trouble with it. I knew it was a great album, but I didn't fully get it yet. It really is great. Very unique part of Devy's catalog.

Back home, ASMR audio for programming, eyes getting heavy.