Wage Elem Improvements
The main one was the styles, making it more streamlined with the rest of the app. Really sleepy, so this may not be my best work. It seems to be okay, with the sleepiness which is falling over me, and drenching my thoughts with nonsense. I am tired, not sure that made any sense. Might make for a good lyric though.

It has a name now because of accessibility. I got a warning for it because I don't check this often enough to know what I missed.